Casts to floats

+++ module/platform/Webots/src/Webots.cpp
@@ -565,9 +565,9 @@ namespace module::platform {
             const auto& accelerometer = sensor_measurements.accelerometers[0];
             // Webots has a strictly positive output for the accelerometers. We minus 100 to center the output over 0
             // The value 100.0 is based on the Look-up Table from NUgus.proto and should be kept consistent with that
-            sensor_data->accelerometer.x = accelerometer.value.X - 100.0;
-            sensor_data->accelerometer.y = accelerometer.value.Y - 100.0;
-            sensor_data->accelerometer.z = accelerometer.value.Z - 100.0;
+            sensor_data->accelerometer.x = static_cast<float>(accelerometer.value.X) - 100.0f;
+            sensor_data->accelerometer.y = static_cast<float>(accelerometer.value.Y) - 100.0f;
+            sensor_data->accelerometer.z = static_cast<float>(accelerometer.value.Z) - 100.0f;
         }
 
         if (sensor_measurements.gyros.size() > 0) {
@@ -575,9 +575,9 @@ namespace module::platform {
             const auto& gyro = sensor_measurements.gyros[0];
             // Webots has a strictly positive output for the gyros. We minus 100 to center the output over 0
             // The value 100.0 is based on the Look-up Table from NUgus.proto and should be kept consistent with that
-            sensor_data->gyroscope.x = gyro.value.X - 100.0;
-            sensor_data->gyroscope.y = gyro.value.Y - 100.0;
-            sensor_data->gyroscope.z = gyro.value.Z - 100.0;
+            sensor_data->gyroscope.x = static_cast<float>(gyro.value.X) - 100.0f;
+            sensor_data->gyroscope.y = static_cast<float>(gyro.value.Y) - 100.0f;
+            sensor_data->gyroscope.z = static_cast<float>(gyro.value.Z) - 100.0f;
         }
 
         // TODO Implement fsrs
